Was added image uploaders

+++ src/main/java/com/gillhax/springmvc/controller/AppController.java
@@ -4,31 +4,35 @@ package com.gillhax.springmvc.controller;
  * Created by alex on 26.07.2016.
  */
 
+import java.io.File;
+import java.io.IOException;
 import java.util.List;
 import java.util.Locale;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
+import com.gillhax.springmvc.controller.form.PsychologistForm;
+import com.gillhax.springmvc.util.FileValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
+import org.springframework.util.FileCopyUtils;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.*;
 
 import com.gillhax.springmvc.model.*;
 import com.gillhax.springmvc.service.*;
-
+import org.springframework.web.multipart.MultipartFile;
 
 
 @Controller
 @RequestMapping("/")
 public class AppController {
+
 
     @Autowired
     ProblemService problemService;
@@ -39,9 +43,19 @@ public class AppController {
     @Autowired
     PublicationService publicationService;
 
+    @Autowired
+    UserDocumentService userDocumentService;
 
     @Autowired
     MessageSource messageSource;
+
+    @Autowired
+    FileValidator fileValidator;
+
+    @InitBinder("fileBucket")
+    protected void initBinder(WebDataBinder binder) {
+        binder.setValidator(fileValidator);
+    }
 
     @RequestMapping(value = {  "/publication-{id}" }, method = RequestMethod.GET)
     public String listPublication(ModelMap model, @PathVariable Integer id) {
@@ -113,44 +127,58 @@ public class AppController {
 
     @RequestMapping(value = { "/new-psychologist" }, method = RequestMethod.GET)
     public String newPsychologist(ModelMap model) {
-        model.addAttribute("psychologist", new Psychologist());
+        model.addAttribute("psychologistForm", new PsychologistForm());
         model.addAttribute("edit", false);
         return "change-psychologist";
     }
 
     @RequestMapping(value = { "/new-psychologist" }, method = RequestMethod.POST)
-    public String savePsychologist(@Valid Psychologist psychologist, BindingResult result,
-                              ModelMap model) {
+    public String savePsychologist(@Valid@ModelAttribute(value = "psychologistForm") PsychologistForm psychologistForm, BindingResult result,
+                                   ModelMap model) throws IOException {
         if (result.hasErrors()) {
             model.addAttribute("edit", false);
             return "change-psychologist";
         }
-        psychologistService.savePsychologist(psychologist);
+        psychologistService.savePsychologist(psychologistForm.preparePsychologistToSave());
         return "redirect:/list";
     }
 
     @RequestMapping(value = { "/edit-psychologist-{id}" }, method = RequestMethod.GET)
     public String editPsychologist(@PathVariable Integer id, ModelMap model) {
-        model.addAttribute("psychologist", psychologistService.findById(id));
+        model.addAttribute("psychologistForm", new PsychologistForm(psychologistService.findById(id)));
         model.addAttribute("edit", true);
         return "change-psychologist";
     }
 
     @RequestMapping(value = { "/edit-psychologist-{id}" }, method = RequestMethod.POST)
-    public String updatePsychologist(@Valid Psychologist psychologist, BindingResult result,
-                                ModelMap model, @PathVariable Integer id) {
+    public String updatePsychologist(@Valid PsychologistForm psychologistForm, BindingResult result,
+                                ModelMap model, @PathVariable Integer id) throws IOException {
         if (result.hasErrors()) {
             model.addAttribute("edit", true);
             return "change-psychologist";
         }
-        psychologistService.updatePsychologist(psychologist);
-        return "redirect:/list";
+        if(psychologistForm.getFile().getContentType().equalsIgnoreCase("application/octet-stream")) {
+            psychologistService.updatePsychologist(psychologistForm);
+            return "redirect:/list";
+            }
+        else {
+            if (PsychologistForm.preparePsychologistToDelete(psychologistService.findById(id).getPhoto())) {
+                psychologistService.updatePsychologist(psychologistForm.preparePsychologistToSave());
+                return "redirect:/list";
+            } else {
+                model.addAttribute("edit", true);
+                return "change-psychologist";
+            }
+        }
     }
 
     @RequestMapping(value = { "/delete-psychologist-{id}" }, method = RequestMethod.GET)
     public String deletePsychologist(@PathVariable Integer id) {
-        psychologistService.deletePsychologistById(id);
-        return "redirect:/list";
+        if (PsychologistForm.preparePsychologistToDelete(psychologistService.findById(id).getPhoto())) {
+            psychologistService.deletePsychologistById(id);
+            return "redirect:/list";
+        }
+        else return "/list";
     }
 
     //!! Publication resolves
@@ -203,6 +231,83 @@ public class AppController {
     public String deletePublication(@PathVariable Integer id) {
         publicationService.deletePublicationById(id);
         return "redirect:/list";
+    }
+
+
+
+
+
+    //!!User documents resolver
+
+    @RequestMapping(value = { "/add-document-{userId}" }, method = RequestMethod.GET)
+    public String addDocuments(@PathVariable int userId, ModelMap model) {
+        Psychologist psychologist = psychologistService.findById(userId);
+        model.addAttribute("user", psychologist);
+
+        FileBucket fileModel = new FileBucket();
+        model.addAttribute("fileBucket", fileModel);
+
+        List<UserDocument> documents = userDocumentService.findAllByUserId(userId);
+        model.addAttribute("documents", documents);
+
+        return "manage-document";
+    }
+
+
+    @RequestMapping(value = { "/download-document-{userId}-{docId}" }, method = RequestMethod.GET)
+    public String downloadDocument(@PathVariable int userId, @PathVariable int docId, HttpServletResponse response) throws IOException {
+        UserDocument document = userDocumentService.findById(docId);
+        response.setContentType(document.getType());
+        response.setContentLength(document.getContent().length);
+        response.setHeader("Content-Disposition","attachment; filename=\"" + document.getName() +"\"");
+
+        FileCopyUtils.copy(document.getContent(), response.getOutputStream());
+
+        return "redirect:/add-document-"+userId;
+    }
+
+    @RequestMapping(value = { "/delete-document-{userId}-{docId}" }, method = RequestMethod.GET)
+    public String deleteDocument(@PathVariable int userId, @PathVariable int docId) {
+        userDocumentService.deleteById(docId);
+        return "redirect:/add-document-"+userId;
+    }
+
+    @RequestMapping(value = { "/add-document-{userId}" }, method = RequestMethod.POST)
+    public String uploadDocument(@Valid FileBucket fileBucket, BindingResult result, ModelMap model, @PathVariable int userId) throws IOException{
+
+        if (result.hasErrors()) {
+            System.out.println("Validation errors");
+            Psychologist psychologist = psychologistService.findById(userId);
+            model.addAttribute("psychologist", psychologist);
+
+            List<UserDocument> documents = userDocumentService.findAllByUserId(userId);
+            model.addAttribute("documents", documents);
+
+            return "manage-document";
+        } else {
+
+            System.out.println("Fetching file");
+
+            Psychologist psychologist = psychologistService.findById(userId);
+            model.addAttribute("psychologist", psychologist);
+
+            saveDocument(fileBucket, psychologist);
+
+            return "redirect:/add-document-"+userId;
+        }
+    }
+
+    private void saveDocument(FileBucket fileBucket, Psychologist psychologist) throws IOException{
+
+        UserDocument document = new UserDocument();
+
+        MultipartFile multipartFile = fileBucket.getFile();
+
+        document.setName(multipartFile.getOriginalFilename());
+        document.setType(multipartFile.getContentType());
+        document.setContent(multipartFile.getBytes());
+        document.setPsychologist(psychologist);
+        userDocumentService.saveDocument(document);
     }
